+++ clang-tools-extra/clang-tidy/misc/ConstCorrectnessCheck.h
@@ -44,6 +44,8 @@ private:
   const bool AnalyzePointers;
   const bool AnalyzeReferences;
   const bool AnalyzeValues;
+  const bool AnalyzeAutoVariables;
+  const bool AnalyzeLambdas;
   const bool AnalyzeParameters;
 
   const bool WarnPointersAsPointers;